+++ packages/crdt_socket_sync/example/server_example/lib/src/registry.dart
@@ -7,8 +7,8 @@ import 'package:hive/hive.dart';
 import 'package:en_logger/en_logger.dart';
 
 const _kDocumentsBox = 'documents';
-const _kDefaultSnapshotInterval = Duration(seconds: 30);
-const _kDefaultMinChangesForSnapshot = 2;
+const _kDefaultSnapshotInterval = Duration(seconds: 30000000);
+const _kDefaultMinChangesForSnapshot = 2000;
 
 /// A server-side CRDT document registry that uses Hive for persistence.
 ///
@@ -232,7 +232,10 @@ class HiveServerRegistry extends CRDTServerRegistry {
     );
 
     await storage.snapshots.saveSnapshot(snapshot);
-    await storage.changes.clear();
+
+    await storage.changes.clear().then((_) {
+      return storage.changes.saveChanges(item.document.exportChanges());
+    });
 
     return snapshot;
   }